130_репозиторий БД в один класс
