+++ app/src/main/java/com/davay/android/feature/matchedsession/di/MatchedSessionFragmentComponent.kt
@@ -8,7 +8,6 @@ import dagger.Component
     dependencies = [AppComponent::class],
     modules = [
         MatchedSessionFragmentModule::class,
-        MatchedSessionDataModule::class,
         MatchedSessionUseCaseModule::class
     ]
 )
