+++ app/src/main/java/com/davay/android/feature/matchedsession/di/MatchedSessionUseCaseModule.kt
@@ -1,6 +1,6 @@
 package com.davay.android.feature.matchedsession.di
 
-import com.davay.android.feature.matchedsession.domain.SessionWithMoviesRepository
+import com.davay.android.core.domain.api.SessionsHistoryRepository
 import com.davay.android.feature.matchedsession.domain.SessionWithMoviesUseCase
 import dagger.Module
 import dagger.Provides
@@ -8,6 +8,6 @@ import dagger.Provides
 @Module
 class MatchedSessionUseCaseModule {
     @Provides
-    fun provideSessionWithMoviesUseCase(sessionWithMoviesRepository: SessionWithMoviesRepository):
-        SessionWithMoviesUseCase = SessionWithMoviesUseCase(sessionWithMoviesRepository)
+    fun provideSessionWithMoviesUseCase(repository: SessionsHistoryRepository):
+        SessionWithMoviesUseCase = SessionWithMoviesUseCase(repository)
 }
