--- app/src/main/java/com/davay/android/feature/matchedsession/domain/SessionWithMoviesRepository.kt
@@ -1,7 +1,0 @@
-package com.davay.android.feature.matchedsession.domain
-
-import com.davay.android.core.domain.models.SessionWithMovies
-
-interface SessionWithMoviesRepository {
-    suspend fun getSessionWithMovies(sessionId: String): SessionWithMovies?
-}
