+++ app/src/main/java/com/davay/android/feature/matchedsession/domain/SessionWithMoviesUseCase.kt
@@ -1,12 +1,13 @@
 package com.davay.android.feature.matchedsession.domain
 
+import com.davay.android.core.domain.api.SessionsHistoryRepository
 import com.davay.android.core.domain.models.SessionWithMovies
 import javax.inject.Inject
 
 class SessionWithMoviesUseCase @Inject constructor(
-    private val sessionWithMoviesRepository: SessionWithMoviesRepository
+    private val repository: SessionsHistoryRepository
 ) {
     suspend fun execute(sessionId: String): SessionWithMovies? {
-        return sessionWithMoviesRepository.getSessionWithMovies(sessionId)
+        return repository.getSessionWithMovies(sessionId)
     }
 }
