--- app/src/main/java/com/davay/android/feature/sessionsmatched/di/MatchedSessionListDataModule.kt
@@ -1,15 +1,0 @@
-package com.davay.android.feature.sessionsmatched.di
-
-import com.davay.android.core.data.database.AppDatabase
-import com.davay.android.feature.sessionsmatched.data.GetSessionsHistoryRepositoryImpl
-import com.davay.android.feature.sessionsmatched.domain.GetSessionsHistoryRepository
-import dagger.Module
-import dagger.Provides
-
-@Module
-class MatchedSessionListDataModule {
-
-    @Provides
-    fun provideGetSessionsHistoryRepository(appDatabase: AppDatabase): GetSessionsHistoryRepository =
-        GetSessionsHistoryRepositoryImpl(appDatabase.historyDao())
-}
