+++ app/src/main/java/com/davay/android/feature/sessionsmatched/di/MatchedSessionListFragmentComponent.kt
@@ -8,7 +8,6 @@ import dagger.Component
     dependencies = [AppComponent::class],
     modules = [
         MatchedSessionListFragmentModule::class,
-        MatchedSessionListDataModule::class,
         MatchedSessionListUseCaseModule::class,
     ]
 )
