+++ app/src/main/java/com/davay/android/feature/sessionsmatched/di/MatchedSessionListUseCaseModule.kt
@@ -1,6 +1,6 @@
 package com.davay.android.feature.sessionsmatched.di
 
-import com.davay.android.feature.sessionsmatched.domain.GetSessionsHistoryRepository
+import com.davay.android.core.domain.api.SessionsHistoryRepository
 import com.davay.android.feature.sessionsmatched.domain.GetSessionsHistoryUseCase
 import dagger.Module
 import dagger.Provides
@@ -8,6 +8,6 @@ import dagger.Provides
 @Module
 class MatchedSessionListUseCaseModule {
     @Provides
-    fun provideGetSessionsHistoryUseCase(getSessionsHistoryRepository: GetSessionsHistoryRepository):
-        GetSessionsHistoryUseCase = GetSessionsHistoryUseCase(getSessionsHistoryRepository)
+    fun provideGetSessionsHistoryUseCase(repository: SessionsHistoryRepository):
+        GetSessionsHistoryUseCase = GetSessionsHistoryUseCase(repository)
 }
