--- app/src/main/java/com/davay/android/feature/sessionsmatched/domain/GetSessionsHistoryRepository.kt
@@ -1,7 +1,0 @@
-package com.davay.android.feature.sessionsmatched.domain
-
-import com.davay.android.core.domain.models.Session
-
-interface GetSessionsHistoryRepository {
-    suspend fun getSessionsHistory(): List<Session>?
-}
